requirements.txt
pyjwt
requests
fire
boto3
flask
flask_httpauth
pyyaml
